requirements.txt
amundsen-databuilder==2.6.2
neo4j-driver==1.7.2
tabulate
pyaml
pyhive
pyinstaller
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
snowflake-connector-python
snowflake-sqlalchemy
